test: Room 레거시 테스트 코드 제거
